+++ bcbio/chipseq/peaks.py
@@ -231,7 +231,7 @@ def consensus(peakfiles, consensusfile, data, pad=250):
 
 def call_consensus(samples):
     """
-    call consensus peaks on the narrow/Broad peakfiles from a set of
+    call consensus peaks on the narrowPeak files from a set of
     ChiP/ATAC samples
     """
     data = samples[0][0]
@@ -243,9 +243,6 @@ def call_consensus(samples):
         if dd.get_chip_method(data) == "chip":
             for fn in tz.get_in(("peaks_files", "macs2"), data, []):
                 if "narrowPeak" in fn:
-                    peakfiles.append(fn)
-                    break
-                elif "broadPeak" in fn:
                     peakfiles.append(fn)
                     break
         elif dd.get_chip_method(data) == "atac":